% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hargeisa}
\alias{hargeisa}
\title{Hargeisa}
\format{
A data frame with 13065 rows and 180 variables:
\describe{
  \item{\code{hargeisa1}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa2}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa3}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa4}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa5}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa6}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa7}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa8}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa9}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa10}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa11}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa12}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa13}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa14}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa15}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa16}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa17}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa18}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa19}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa20}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa21}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa22}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa23}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa24}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa25}}{logical COLUMN_DESCRIPTION}
  \item{\code{hargeisa26}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa27}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa28}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa29}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa30}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa31}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa32}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa33}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa34}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa35}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa36}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa37}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa38}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa39}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa40}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa41}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa42}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa43}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa44}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa45}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa46}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa47}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa48}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa49}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa50}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa51}}{logical COLUMN_DESCRIPTION}
  \item{\code{hargeisa52}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa53}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa54}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa55}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa56}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa57}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa58}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa59}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa60}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa61}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa62}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa63}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa64}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa65}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa66}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa67}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa68}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa69}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa70}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa71}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa72}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa73}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa74}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa75}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa76}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa77}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa78}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa79}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa80}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa81}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa82}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa83}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa84}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa85}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa86}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa87}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa88}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa89}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa90}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa91}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa92}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa93}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa94}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa95}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa96}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa97}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa98}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa99}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa100}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa101}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa102}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa103}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa104}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa105}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa106}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa107}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa108}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa109}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa110}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa111}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa112}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa113}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa114}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa115}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa116}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa117}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa118}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa119}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa120}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa121}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa122}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa123}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa124}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa125}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa126}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa127}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa128}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa129}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa130}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa131}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa132}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa133}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa134}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa135}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa136}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa137}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa138}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa139}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa140}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa141}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa142}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa143}}{logical COLUMN_DESCRIPTION}
  \item{\code{hargeisa144}}{logical COLUMN_DESCRIPTION}
  \item{\code{hargeisa145}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa146}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa147}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa148}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa149}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa150}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa151}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa152}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa153}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa154}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa155}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa156}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa157}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa158}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa159}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa160}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa161}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa162}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa163}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa164}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa165}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa166}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa167}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa168}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa169}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa170}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa171}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa172}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa173}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa174}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa175}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa176}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa177}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa178}}{character COLUMN_DESCRIPTION}
  \item{\code{hargeisa179}}{double COLUMN_DESCRIPTION}
  \item{\code{hargeisa180}}{double COLUMN_DESCRIPTION} 
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
hargeisa
}
\description{
Hargeisa
}
\keyword{datasets}
