% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{use_cells}
\alias{use_cells}
\title{use_cells}
\usage{
use_cells(x, y, data, benchmark, combination_cells, combination_indicators)
}
\arguments{
\item{\dots}{List of packages}
}
\description{
Use population cells
}
