% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{use_criterion}
\alias{use_criterion}
\title{use_criterion}
\usage{
use_criterion(data, sim_data, benchmark, x)
}
\arguments{
\item{\dots}{List of packages}
}
\description{
Composite measure at criterion level
}
