% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{use_subcriterion}
\alias{use_subcriterion}
\title{use_subcriterion}
\usage{
use_subcriterion(data, sim_data, benchmark, x)
}
\arguments{
\item{\dots}{List of packages}
}
\description{
Composite measure at subcriterion level
}
